manage jinja templates as nicely formatted files
