+++ src/axolotl/utils/chat_templates/templates/alpaca.jinja
@@ -1,0 +1,21 @@
+{{ bos_token }}
+{% for message in messages %}
+    {% if message['role'] == 'system' and loop.first %}
+        {{ message['content'] }}
+    {% elif message['role'] == 'user' %}
+        {{ '### Instruction:
+        ' + message['content'] }}
+    {% elif message['role'] == 'assistant' %}
+        {{ '### Response:
+        ' + message['content'] + eos_token }}
+    {% endif %}
+    {% if not loop.last %}
+        {{ '
+        ' }}
+    {% endif %}
+{% endfor %}
+{% if add_generation_prompt %}
+    {{ '
+    ### Response:
+    ' }}
+{% endif %}
