+++ src/axolotl/utils/chat_templates/templates/chatml.jinja
@@ -1,0 +1,12 @@
+{% if not add_generation_prompt is defined %}
+    {% set add_generation_prompt = false %}
+{% endif %}
+{% for message in messages %}
+    {{'<|im_start |>' + message['role'] + '
+    ' + message['content'] + '<|im_end |>' + '
+    '}}
+{% endfor %}
+{% if add_generation_prompt %}
+    {{ '<|im_start |>assistant
+    ' }}
+{% endif %}
