+++ src/axolotl/utils/chat_templates/templates/deepseek_v2.jinja
@@ -1,0 +1,15 @@
+{% if not add_generation_prompt is defined %}
+    {% set add_generation_prompt = false %}
+{% endif %}
+{{ bos_token }}
+{% for message in messages %}
+    {% if message['role'] == 'user' %}
+        {{ '<｜User ｜>' + message['content'] }}
+    {% elif message['role'] == 'assistant' %}
+        {{ '<｜Assistant ｜>' + message['content'] + eos_token }}
+    {% elif message['role'] == 'system' %}
+        {{ message['content'] + '
+        ' }}
+    {% endif %}
+{% endfor %}
+{% if add_generation_prompt %}{{ '<｜Assistant ｜>' }}{% endif %}
