+++ src/axolotl/utils/chat_templates/templates/exaone.jinja
@@ -1,0 +1,15 @@
+{% for message in messages %}
+    {% if loop.first and message['role'] != 'system' %}
+        {{ '[|system|][|endofturn|]
+        ' }}
+    {% endif %}
+    {{ '[|' + message['role'] + '|]' + message['content'] }}
+    {% if message['role'] == 'user' %}
+        {{ '
+        ' }}
+    {% else %}
+        {{ '[|endofturn|]
+        ' }}
+    {% endif %}
+{% endfor %}
+{% if add_generation_prompt %}{{ '[|assistant|]' }}{% endif %}
