+++ src/axolotl/utils/chat_templates/templates/llama3.jinja
@@ -1,0 +1,16 @@
+{% if not add_generation_prompt is defined %}
+    {% set add_generation_prompt = false %}
+{% endif %}
+{% set loop_messages = messages %}
+{% for message in loop_messages %}
+    {% set content = '<|start_header_id |>' + message['role'] + '<|end_header_id |>
+        '+ message['content'] | trim + '<|eot_id |>' %}
+    {% if loop.index0 == 0 %}
+        {% set content = bos_token + content %}
+    {% endif %}
+    {{ content }}
+{% endfor %}
+{% if add_generation_prompt %}
+    {{ '<|start_header_id |>assistant<|end_header_id |>
+    ' }}
+{% endif %}
