+++ src/axolotl/utils/chat_templates/templates/metharme.jinja
@@ -1,0 +1,20 @@
+{{ bos_token }}
+{% if messages[0]['role'] == 'system' %}
+    {% set loop_messages = messages[1:] %}{% set system_message = messages[0]['content'] %}
+{% else %}
+    {% set loop_messages = messages %}{% set system_message = "Enter RP mode. You shall reply to the user while staying in character. Your responses must be detailed, creative, immersive, and drive the scenario forward." %}
+{% endif %}
+{{ '<|system |>' + system_message }}
+{% for message in loop_messages %}
+    {% set content = message['content'] %}
+    {% if message['role'] == 'user' %}
+        {{ '<|user |>' + content.strip() }}
+    {% elif message['role'] == 'assistant' %}
+        {{ '<|model |>'  + content.strip() }}
+    {% endif %}
+{% endfor %}
+{% if add_generation_prompt %}
+    {{ '<|model |>' }}
+{% else %}
+    {{ eos_token }}
+{% endif %}
