+++ src/axolotl/utils/chat_templates/templates/mistral_v3_tekken.jinja
@@ -1,0 +1,13 @@
+{{ bos_token }}
+{% for message in messages %}
+    {% if (message['role'] == 'user') != (loop.index0 % 2 == 0) %}
+        {{ raise_exception("Conversation roles must alternate user/assistant/user/assistant/...") }}
+    {% endif %}
+    {% if message['role'] == 'user' %}
+        {{ '[INST]' + message['content'] + '[/INST]' }}
+    {% elif message['role'] == 'assistant' %}
+        {{ message['content'] + eos_token }}
+    {% else %}
+        {{ raise_exception("Only user and assistant roles are supported!") }}
+    {% endif %}
+{% endfor %}
