+++ src/axolotl/utils/chat_templates/templates/phi_3.jinja
@@ -1,0 +1,16 @@
+{{ bos_token }}
+{% for message in messages %}
+    {% if (message['role'] == 'system') %}
+        {{'<|system |>' + '
+        ' + message['content'] + '<|end |>' + '
+        '}}
+    {% elif (message['role'] == 'user') %}
+        {{'<|user |>' + '
+        ' + message['content'] + '<|end |>' + '
+        ' + '<|assistant |>' + '
+        '}}
+    {% elif message['role'] == 'assistant' %}
+        {{message['content'] + '<|end |>' + '
+        '}}
+    {% endif %}
+{% endfor %}
