+++ src/axolotl/utils/chat_templates/templates/phi_35.jinja
@@ -1,0 +1,19 @@
+{% for message in messages %}
+    {% if message['role'] == 'system' and message['content'] %}
+        {{'<|system |>
+        ' + message['content'] + '<|end |>
+        '}}
+    {% elif message['role'] == 'user' %}
+        {{'<|user |>
+        ' + message['content'] + '<|end |>
+        '}}
+    {% elif message['role'] == 'assistant' %}
+        {{'<|assistant |>
+        ' + message['content'] + '<|end |>
+        '}}
+    {% endif %}
+{% endfor %}
+{% if add_generation_prompt %}
+    {{ '<|assistant |>
+    ' }}
+{% endif %}
